Move contact loading logic into the new ContactPickerViewModel.kt This also loads contact photos once, during contact loading
